requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
langchain==0.1.0
langchain-community==0.0.13
langchain-groq==0.0.1
sentence-transformers==2.3.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
huggingface-hub==0.20.2
